add changes to build and push to docker hub

+++ travis/tests-unit.sh
@@ -1,5 +1,10 @@
 set -e
 
+echo ""
+echo "UPSTREAM_BRANCH: $UPSTREAM_BRANCH"
+echo "UPSTREAM_TAG: $UPSTREAM_TAG"
+echo ""
+
 if [ "$TEST_EASY_INSTALL" == "true" ]; then
     cp ./easy-deploy-script/easy-deploy-viral-ngs.sh /tmp/easy-deploy-viral-ngs.sh
     /tmp/easy-deploy-viral-ngs.sh setup
@@ -10,9 +15,19 @@ if [ "$TEST_ANSIBLE" == "true" ]; then
 fi
 
 if [ "$TEST_DOCKER" == "true" ]; then
+    export REPO=broadinstitute/viral-ngs
+    export TAG=`if [ "$UPSTREAM_TAG" == "master" ]; then echo "latest"; else echo $UPSTREAM_TAG ; fi`
     # tar contents to dereference symlinks
     cd ./docker
-    # build the docker image, and try to run it
-    tar -czh . | docker build --rm -q - | xargs -I{} docker run --rm {} illumina.py
+
+    if [ -z "$UPSTREAM_TAG" ]; then
+        # build the docker image, and try to run it
+        tar -czh . | docker build --rm -q - | xargs -I{} docker run --rm {} illumina.py
+    # if this was triggered by the upstream repo, build, tag, and push to Docker Hub
+    else
+        docker login -e $DOCKER_EMAIL -u $DOCKER_USER -p $DOCKER_PASS
+        tar -czh . | docker build --rm -t "$REPO:$TAG" -q - | xargs -I{} docker run --rm {} illumina.py && docker push "$REPO:$TAG"
+    else
+    fi
 fi
